+++ docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+* Added package `ml/data/downloader` for parallel downloads, with support for authentication tokens.
+* Added package `ml/data/huggingface` to download HuggingFace models.
+* Removed dependency to gonb/common. Added package `types/xsync` with the required synchronization constructs.
 * Added `Shape.Dim(axis)` as a shortcut, where `axis` can use negative values.
 * Package `graph`:
   * `Scalar()`, `AddScalar()`, `MulScalar()`, `DivScalar()`, ... are now generic, and take as input any non-complex